Fix thing

+++ backend/compilers/download.py
@@ -990,7 +990,8 @@ def download_win9x():
         dest_name="msvc40",
     )
     shutil.move(COMPILERS_DIR / "MSVC400-master", COMPILERS_DIR / "msvc4.0")
-    shutil.move(COMPILERS_DIR / "msvc4.0/BIN", COMPILERS_DIR / "msvc4.0/Bin")
+    if not os.path.exists(COMPILERS_DIR / "msvc4.0/BIN"):
+        shutil.move(COMPILERS_DIR / "msvc4.0/BIN", COMPILERS_DIR / "msvc4.0/Bin")
     set_x(COMPILERS_DIR / "msvc4.0/Bin/CL.EXE")
 
     download_zip(
@@ -998,7 +999,8 @@ def download_win9x():
         dest_name="msvc42",
     )
     shutil.move(COMPILERS_DIR / "MSVC420-master", COMPILERS_DIR / "msvc4.2")
-    shutil.move(COMPILERS_DIR / "msvc4.2/bin", COMPILERS_DIR / "msvc4.2/Bin")
+    if not os.path.exists(COMPILERS_DIR / "msvc4.2/Bin"):
+        shutil.move(COMPILERS_DIR / "msvc4.2/bin", COMPILERS_DIR / "msvc4.2/Bin")
     set_x(COMPILERS_DIR / "msvc4.2/Bin/CL.EXE")
